:white_check_mark: new bookmark tests
