+++ test/constants/selectors.js
@@ -26,6 +26,7 @@ const ss = {
   voterAddress: '.voters .voter-address',
   accountAddress: '.copy-title',
   selectionVotingNumber: '.selection h4',
+  totalVotingNumber: '.total h4',
   searchInput: '#autosuggest-input',
   showMoreVotesBtn: '.show-votes',
   leftBlockAccountExplorer: '.explorer-account-left-block',
@@ -119,7 +120,7 @@ const ss = {
   addressInput: '.address input',
   followedAccountItem: '.followed-account',
   titleInput: '.account-title input',
-  followedAccountBalance: '.followed-account-balance',
+  followedAccountBalance: '.followed-account-balance span',
   sidebarMenuHelpBtn: '#help',
   startOnBoardingLink: '.help-onboarding',
   onBoardingTooltipPrimaryBtn: '.joyride-tooltip__button--primary',
