+++ test/constants/urls.js
@@ -12,6 +12,7 @@ const urls = {
   delegates: '/delegates',
   delegatesVote: '/delegates/vote',
   signMessage: '/sign-message',
+  sidechains: '/sidechains',
 };
 
 export default urls;
