+++ test/cypress/e2e/bookmarks.spec.js
@@ -1,69 +1,103 @@
 import accounts from '../../constants/accounts';
 import networks from '../../constants/networks';
-import ss from '../../constants/selectors';
 import urls from '../../constants/urls';
+import ss from '../../constants/selectors';
+import compareBalances from '../utils/compareBalances';
+
+const getFollowedAccountObjFromLS = () => JSON.parse(localStorage.getItem('followedAccounts'));
+
+const txConfirmationTimeout = 14000;
 
 describe('Bookmarks', () => {
   /**
-   * Bookmarks suggestions are not present if there is no followers
-   * @expect bookmarks components are not present
+   * Add follower in dashboard widget
+   * @expect localStorage have the follower object with correct address
+   * @expect localStorage have the follower object with correct title
+   * @expect following account is shown in widget
    */
-  it('Bookmarks are not present if there is no followers', () => {
+  it('Add bookmark while logged in', () => {
     cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit(urls.send);
-    cy.get(ss.recipientInput).click();
-    cy.get(ss.bookmarkInput).should('not.exist');
-    cy.get(ss.bookmarkList).should('not.exist');
+    cy.visit(urls.dashboard);
+    cy.get(ss.addFollowerAccountButton).click();
+    cy.get(ss.addressInput).click().type(accounts.genesis.address);
+    cy.get(ss.nextBtn).click();
+    cy.get(ss.titleInput).type('Bob');
+    cy.get(ss.nextBtn).click()
+      .should(() => {
+        expect(getFollowedAccountObjFromLS()[0].address).to.equal(accounts.genesis.address);
+        expect(getFollowedAccountObjFromLS()[0].title).to.equal('Bob');
+      });
+    cy.get(ss.followedAccountItem).should('have.length', 1);
+    cy.get(ss.titleInput).should('have.value', 'Bob');
   });
 
-  /**
-   * Choose follower from bookmarks and send tx
-   * @expect bookmark contain name
-   * @expect bookmark contain address
-   * @expect clicking bookmark fills recipient
-   * @expect tx appears in activity with right address
-   */
-  it('Choose follower from bookmarks and send tx', () => {
-    window.localStorage.setItem('followedAccounts', `[
-      {"title":"Alice","address":"${accounts.delegate.address}","balance":101}
-    ]`);
-    cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit(urls.send);
-    cy.get(ss.recipientInput).click();
-    cy.get(ss.bookmarkInput);
-    cy.get(ss.bookmarkList).eq(0).contains('Alice');
-    cy.get(ss.bookmarkList).eq(0).contains(accounts.delegate.address);
-    cy.get(ss.bookmarkList).eq(0).click();
-    cy.get(ss.recipientInput).should('have.value', accounts.delegate.address);
-    cy.get(ss.amountInput).click().type(1);
-    cy.get(ss.nextTransferBtn).click();
-    cy.get(ss.sendBtn).click();
-    cy.get(ss.okayBtn).click();
-    cy.get(ss.transactionRow).eq(0).find(ss.transactionAddress).should('have.text', accounts.delegate.address);
-  });
+  describe('Saved bookmark in localStorage', () => {
+    beforeEach(() => {
+      window.localStorage.setItem('followedAccounts', `[{"title":"Alice","address":"${accounts.genesis.address}","balance":101}]`);
+    });
 
-  /**
-   * Search through bookmarks by typing
-   * @expect account found by name
-   * @expect non-existent search show empty list
-   * @expect account found by delegate
-   */
-  it('Search through bookmarks by typing', () => {
-    window.localStorage.setItem('followedAccounts', `[
-      {"title":"Alice","address":"${accounts.delegate.address}","balance":101},
-      {"title":"Bob","address":"${accounts.genesis.address}","balance":101}
-    ]`);
-    cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
-    cy.visit(urls.send);
-    cy.get(ss.recipientInput).click().type('Bob');
-    cy.get(ss.bookmarkList).eq(0).contains('Bob');
-    cy.get(ss.recipientInput).clear();
-    cy.get(ss.recipientInput).click().type('Merkel');
-    cy.get(ss.bookmarkList).should('not.exist');
-    cy.get(ss.recipientInput).clear();
-    cy.get(ss.recipientInput).click().type(accounts.delegate.address);
-    cy.get(ss.bookmarkList).eq(0).contains(accounts.delegate.address);
-    cy.get(ss.bookmarkList).eq(0).click();
-    cy.get(ss.recipientInput).should('have.value', accounts.delegate.address);
+    /**
+     * Edit followers title
+     * @expect value has changed
+     */
+    it('Edit bookmarks title', () => {
+      cy.visit(urls.dashboard);
+      cy.get(ss.editFollowingAccounts).click();
+      cy.get(ss.titleInput).clear().type('Bob');
+      cy.get(ss.editFollowingAccounts).click();
+      cy.get(ss.followedAccountTitle).should('have.value', 'Bob');
+    });
+
+    /**
+     * Delete a follower
+     * @expect follower is not there
+     */
+    it('Delete a bookmark', () => {
+      cy.visit(urls.dashboard);
+      cy.get(ss.editFollowingAccounts).click();
+      cy.get(ss.removeFollowingAccount).click();
+      cy.get(ss.followedAccount).should('not.exist');
+    });
+
+    /**
+     * Follower account is clickable
+     * @expect follower account is opened
+     */
+    it('Open bookmarked account', () => {
+      cy.visit(urls.dashboard);
+      cy.get(ss.followedAccount).click();
+      cy.url().should('include', '/explorer/accounts/16313739661670634666L');
+    });
+
+    /**
+     * Try to add duplicate followed account
+     * @expect button is disabled
+     */
+    it('Not possible to add duplicate', () => {
+      cy.visit(urls.dashboard);
+      cy.get(ss.addFollowerAccountButton).click();
+      cy.get(ss.addressInput).click().type(accounts.genesis.address);
+      cy.get(ss.nextBtn).should('be.disabled');
+    });
+
+    it('Balance is correct and get live updates', function () {
+      window.localStorage.setItem('followedAccounts', `[
+        {"title":"Alice","address":"${accounts.genesis.address}","balance":101}
+      ]`);
+      cy.autologin(accounts.genesis.passphrase, networks.devnet.node);
+      cy.visit(urls.dashboard);
+      cy.get(ss.followedAccountBalance).invoke('text').as('balanceBefore');
+      cy.visit(urls.send);
+      cy.get(ss.recipientInput).type(accounts.genesis.address);
+      cy.get(ss.amountInput).click().type('1');
+      cy.get(ss.nextTransferBtn).click();
+      cy.get(ss.sendBtn).click();
+      cy.visit(urls.dashboard);
+      cy.get(ss.transactionRow).eq(0).find(ss.spinner);
+      cy.get(ss.transactionRow).eq(0).find(ss.spinner, { timeout: txConfirmationTimeout }).should('not.exist');
+      cy.get(ss.followedAccountBalance).invoke('text').as('balanceAfter').then(() => {
+        compareBalances(this.balanceBefore, this.balanceAfter, 0.1);
+      });
+    });
   });
 });
